# GPKG: use much faster creation of RTree with a in-memory RTree building