--- a/ogr/ogrsf_frmts/sqlite/sqlite_rtree_bulk_load/README
+++ b/ogr/ogrsf_frmts/sqlite/sqlite_rtree_bulk_load/README
@@ -1,0 +1,1 @@
+Upstream is at https://github.com/rouault/sqlite_rtree_bulk_load