--- a/ogr/ogrsf_frmts/sqlite/sqlite_rtree_bulk_load/wrapper.cpp
+++ b/ogr/ogrsf_frmts/sqlite/sqlite_rtree_bulk_load/wrapper.cpp
@@ -1,0 +1,3 @@
+#define SQLITE_RTREE_BL_SYMBOL(x) gdal_##x
+
+#include "sqlite_rtree_bulk_load.c"