… ogr/ogrsf_frmts/sqlite/CMakeLists.txt
… ogr/ogrsf_frmts/sqlite/sqlite_rtree_bulk_load/sqlite_rtree_bulk_load.c
… ogr/ogrsf_frmts/sqlite/sqlite_rtree_bulk_load/sqlite_rtree_bulk_load.h
… scripts/fix_typos.sh