With optimization:
```
$ time ogr2ogr tmp.gpkg nz-building-outlines.gpkg
real    0m16,433s
user    0m18,093s
sys     0m2,135s
```

vs without optimization with sqlite 3.43.2 (modified to integrate the disabling of RTree forced reinsertion from SQLite master https://github.com/sqlite/sqlite/commit/7de8ae22f7f895d5564f2f006d22e2dfddfdf5a7. Otherwise with stock 3.43.2, it is 38 seconds)

```
$ time OGR_GPKG_MAX_RAM_USAGE_RTREE=0 ogr2ogr tmp.gpkg nz-building-outlines.gpkg
real    0m28,000s
user    0m40,287s
sys     0m5,244s
```

Fixes #7614

The core functionality is available as a library in https://github.com/rouault/sqlite_rtree_bulk_load